+++ test/merge_types.test.js
@@ -1,3 +1,5 @@
+import { parse } from 'graphql';
+
 import mergeTypes from '../src/merge_types';
 import clientType from './graphql/types/client_type';
 import productType from './graphql/types/product_type';
@@ -481,6 +483,23 @@ describe('mergeTypes', () => {
 
   it('preserves the input field comments', () => {
     const types = [clientType, productType];
+    const mergedTypes = mergeTypes(types);
+    const expectedClientType = normalizeWhitespace(`
+      input ClientFormInputWithComment {
+        # Name
+        name: String!
+        # Age
+        age: Int!
+      }
+    `);
+    const separateTypes = normalizeWhitespace(mergedTypes);
+
+    expect(separateTypes).toContain(expectedClientType);
+  });
+
+  it('supports already parsed documents', () => {
+    const parsedClientType = parse(clientType);
+    const types = [parsedClientType, productType];
     const mergedTypes = mergeTypes(types);
     const expectedClientType = normalizeWhitespace(`
       input ClientFormInputWithComment {