+++ test/polyfills.js
@@ -1,0 +1,1 @@
+Object.values = obj => Object.keys(obj).map(key => obj[key]);